return specific column names in update/insert queries

+++ src/db/queries/account/role/drop-all-user-roles.sql
@@ -1,2 +1,2 @@
 
-DELETE FROM account.user_role WHERE user_id=$1 RETURNING *;
+DELETE FROM account.user_role WHERE user_id=$1 RETURNING user_id, role_id;
